hieradata: add alternate slurm state save dir loc
Default Slurm state save location is /var/spool and its not OK. Then we
define a new default value globally in hiera. The value has not been
set in the params class of the slurm module because it might usefull to
know this location for profiles and other hiera parameters, typically
to mount a shared filesystem between slurmctld server. We also define a
shared parameter slurm_state_save_loc in hiera for this purpose.