mod:slurm merge slurm::client into slurm
The slurm::client classes installed the slurm clients. Those clients
are needed everywhere the slurm configuration file is deployed. Then,
there was no reason to separate slurm and slurm::client. This commit
merges slurm::client and slurm into the slurm classes.

The sview graphical utility is not installed anymore by default because
it simply sucks. If people want to install it, they must give it into
the slurm::packages array.