first draft of new doc
This commit introduces the new draft of Puppet-HPC documentation plan.

The previous documentation has been moved temporarily into doc/old
subdirectory but the latex template has been removed. The new
documentation building process depends on edf-doc-materials package
which provides most pieces required to build various formats easily.

This first draft is basically a collection of commented-out TODOs, to
be replaced and completed incrementaly.

There is also a new script gen-ref.rb to generate the reference API
section of the documentation based on documentation snippets directly
written in puppet code. For the moment, the script generates an empty
file. To be developped as well.