hieradata: set fqdn in slurm mysql setup conf
The internal DNS server response with the local_domain as domain.
Because the domain of nodes is not the local_domain but the domain in
hiera, mysql consider the full FQDN of client hosts when resolving
reverse DNS of incoming connections and when checking against its
grants configuration. Then, the grants must be configured with nodes
FQDN with the local_domain as domain to match responses from the
reverse DNS.