mod:hpclib create reverse zone for each /24 subnet
To avoid having a too big reverse zone that could potentially overlap
external subnets, the algorithm creates a reverse zone for every /24
subnet of the cluster networks.

The current implementation of the algorithm sucks since it works only
for networks having CIDR prefixes between 16 and 24. It really has to
be improved somehow or consider more advanced things like RFC 2317.