mod:boothttp: transform supported_os to array
The supported_os parameter was a hash with keys equivalent to its
content. This was done on purpose in order to use create_resources()
which supported hash only. Now, hpc_atoh() is used to transform the
array into a hash and the printconfig() defined type use the resource
name as the os flavor.