mod:dns: manage virtual zone if relay if false
The dns::server public class is now able to manage the virtual zone if
the virtual_relay parameter is false. The zone contains the given
virtual_entries. This is useful in the early stage of the cluster
installation process to get the services DNS names when consul is not
ready yet. Documentation of the module has been updated accordingly.