prof:jobsched ceph StateSave not mandatory
The reference configuration for scibian HPC cluster uses a cephFS
mount for the StateSaveLocation between the slurmctld.

The boolean introduced here disable the ceph mount if the site
deviate from this configuration (for tests purposes). It is set
to true by default in the common hiera (previous behavior by
default).