mod:multipath: add support for subsections in conf
Manage subsections inside sections of the multipath.conf file.

The multipath subsections (inside the multipaths section) have to be
managed separately in a paths array because of the conflicting
'multipath' keys in the hash. When defining multiple multipath keys in
the input hash, only one is present in the resulting configuration file.
This is workaround using this additional array.