prof:http reject all traffic not explicitely authorized to secret
This commit removes a security vulnerability.

The previous version of the rules relied on a REJECT default policy
for traffic coming from the outside of the cluster. This is not
allways true.

The new version of the rules explicitely creates a default rule for
this port. It also introduces two exceptions that where implicit
before: localhost (fw zone) and from cluster (clstr zone) with
a source port strictly below 1024.

The localhost must be removed eventually in case a user has access
to a service node. It is maintained temporarily until the consul
check is fixed.