mod:nfs Implement default_options, nfsd threads=64
With the NFS server, some settings are set by a default/sysconfig
file. This commit handle this by introducing a hash of values to
put in this file. This commit uses this to change the default to
64 threads for the nfsd server.

The default of 8 threads is not enough for the large I/Os that
NFS servers on large cluster must handle.