mod:network: use hash for routednet
The routednet parameter was an array on triplets strings. With this
commit it becomes a hash (netifs are keys) of array of networks. This
notably simplifies the interfaces templates since there is no triplet
logic to run anymore. The commit also add an example of this parameter
in README.md documentation file.

Fix #96