mod:network: use static mode for netifs
Since the beginning of Puppet-HPC, the network interfaces on Debian were
setup in manual mode with many pre-up/up statements to configure
everything from IP addresses to IB tuning. I think this is the root
cause of many issues in the boot process, such as services startup fail
because depending on NFS share over a bonding interface. Anyway, the
network interfaces with IP addresses are now setup in static mode, the
standard ifupdown way to proceed. I hope this is going to improve those
issues in the long term. At least, it largely improves readability of
/etc/network/interfaces file.

With this commit, the interfaces are also setup with auto/allow-auto
standard tag to configure network interfaces at boottime with networking
service. Initially, the module setup interfaces with the allow-hotplug
tag and an additional ifup-hotplug service. This change makes this
service obsolete as default networking service do all the work properly.
Therefore, this commit also removes all reference to the ifup-hotplug
service. On installed nodes, the cleanup of this service must be
performed manually.

The commit also provides if-pre-up.d part script to kill running
instances of dhclient and setup IB/OPA interfaces modes. This change
makes able to remove weird pre-up statement commands.

Please note that this commit implies a complete rewrite of the
/etc/network/interfaces files and networking service restart. Therefore
it may break your network setup. Make sure you have access to the node
remote serial console before running this code.

Fix #70