mod:slurm: manage slurm spank plugins conf
The slurm module is now able to optionally manage the configuration
files of spank plugins. The module does not hardcode a default list of
spank plugins to install anymore as it is something that is rather site
specific. The plugins with their optional configuration are injected
into slurm module with the new spank_plugins hash. The usage of this
parameter is documented in the README.md file with an example.

Fix #88