mod:network Test network is actually up after service start
When booting diskless nodes, the first apt-get install was failing
because networking restart had returned but the interfaces were
not actually up. An exec is added to not return until the gateway
actually pings, this small delay is enough.

This ping uses fping and is added as a utils package.