mod:infiniband: remove mxm package
This package install files in /opt directory which is a shared
filesystem on diskless compute nodes of EDF clusters. For the moment,
puppet MUST NOT install this package on compute nodes as it would badly
interfere with the content of the shared filesystem. The mxm package
is cared by the cluster's meta-packages (like most other infiniband
packages as well) and it is fine this way.