mod:pam Realize all pam resources after all pam-auth-update have run
The pam files can be modified by pam-auth-update executed directly
or by a package install. The direct modification of the pam files
by the pam resources does not mix well. This patch make sure all
the pam resources are executed on stable files.

It also make sure that the entry if it is missing is added in the
right block (pam-auth-update or not)