prof:consul: fix ceph definition test
The result of build-in Puppet function defined(<resource>) depends on
evaluation order. The value can be false if the resource is realized
later in the evaluation order. This is fully explained in Puppet
documentation:

  https://docs.puppet.com/puppet/latest/function.html#defined

A more reliable way to test this is to check if the node is member of
the array of nodes having the ceph::server profile.

Also change merge() for concat() which is designed for arrays while the
first is for hashes.