hiera Remove KexAlgorithms and AcceptEnv from common.yaml
Those settings only make sense on frontend and frontend/nodes
respectively. These settings are also hard to write in a portable
way between RHEL and Debian.

Removing these from the common level is cleaner and easier even if
it will mean duplication at the role level.

breaks: Settings have to be moved to the role level, frontend for
X11UseLocalhost, ClientAliveInterval and KexAlgorithms. All frontend
and compute nodes for AcceptEnv.