mod:network Support for RHEL
Add support for RHEL.

Only features not reimplemented is eth_no_offload_ifs.

This change is backward compatible with existing Debian configuration.

The main change is that on RHEL there is a configuration file per
interface, so the net_ifaces had to be completed with implicit
interfaces from the puppet code. On Debian those are discovered inside
the template itself.