mod:gpfs: add lum files workaround
On RHEL, the RPM packages of GPFS < 4.2 provided by IBM do provide the
lum blobs (licensing stuff) in PTF-0 releases of the packages. The
installation of packages in version PTF >0 from scratch systematically
fail due to missing lum blobs. To install these packages, the blobs must
be provided through another way. Those new parameter are ugly but it
does the workaround.

These parameters are supposed to be used with hiera autolookups in
specific roles only. This is ugly but it is fixed with GPFS (aka.
Spectrum Scale) >= 4.2, then it will be possible to remove this
crazyness from the module anytime soon, hopefully.