prof:gpfs: extract ssl_keys hash from hiera
The ssl_keys parameter for profiles::gpfs::clusters hash is now a hash
of SSL keys definitions (with a source and an optional filename). The
profile now gives this new parameter to the gpfs public class.

Breaks: this change requires to update profiles::gpfs::clusters hash in
hiera repositories in order to replace key_src string parameter with new
ssl_keys hash. The optional key_file parameter has been removed. The
hiera repositories can also be cleaned-up if it was used.