hiera: ntp restrict and interface
By default the profiles ntp::client and ntp::server will be restricted
this way:

client: listen only on localhost and administration, replies only
to localhost.

server: listen only on localhost, wan, management and administration,
replies only on localhost, management and administration.

NTP must still listen on the network to the upstream server because
it can send a request on a network it don't listen to.

breaks: Relies on some keys from network.yaml:
 - net::{wan,administration,management}::ipnetwork
 - net::{wan,administration,management}::prefix_length