prof:http Lock down http:secret further
This rework of the secret vhost makes it listen only on localhost and
redirect add firewall to redirect remote traffic to localhost.

This new scheme makes the vhost more secure while the firewall is down, because
the vhost is not accessible remotely without the proper firewall configuration.

This update also turns the action for outside packet to DROP to match the new
default policy. The port no longer stands out as being "closed".

Notice this enable route_localnet to make DNAT to localhost work.