mod:slurmutils: add opt importers to mysql-setup
This commit adds an additional importers array parameter to
slurmutils::setupdb in order to optionally restrict the list of source
host for the slurmro DB user. When not defined, this parameter default
to undef and is not present in configuration file. This makes the
mysql-setup utility create a slurmro@% user granted for any host.