mod:icinga2: add param to bind API socket
This commit introduces a new bind_host parameter to the icinga2 public
class in order to optionally bind the API listener socket to a specific
IP address (or hostname). If not set, the parameter default to undef
which makes the public class to not set the parameter in API listener
configuration.