prof:apt proxy: Also search VIPs for listen_networks
apt.service.virtual can refer to the VIP of the service node during
very early boot of diskless nodes.

This patch add all VIPs of the node to the listen interfaces. Since
some of those addresses might not be available during system start,
this patch also sets the sysctl net.ipv4.ip_nonlocal_bind to
authorize binding an interface that does not exists.