prof:bootsystem server: Add VIPs of listen_networks for http
During boot, the node might use the VIP to connect to the boot
system instead of the dns name provided by consul.

This patch add all VIPs of the node to the listen interfaces. Since
some of those addresses might not be available during system start,
this patch also sets the sysctl net.ipv4.ip_nonlocal_bind to
authorize binding an interface that does not exists.