prof:http diskless: Add support for listen_networks parameter
The listen_networks hiera key permits to set a lis of networks that
will be used by the vhost to listen to.

This patch add all VIPs of the node to the listen interfaces. Since
some of those addresses might not be available during system start,
this patch also sets the sysctl net.ipv4.ip_nonlocal_bind to
authorize binding an interface that does not exists.