prof:inventory collect: Add listen_networks parameter for vhost
For security purpose it is preferable to restrict the listen interface
of the vhost to the ones that will actually be used.

This patch add all VIPs of the node to the listen interfaces, since
some of those addresses might not be available during system start,
this patch also sets the sysctl net.ipv4.ip_nonlocal_bind to
authorize binding an interface that does not exists.