mod:systemd Add support for other styles in unit_override and restart
When doing an override on parameter that can be specified multiple
times, it is usefull to have this kind of construct:

Option=
Option=valueA
Option=valueB

The ini style in print_config does not handle this case because it
is a hash. This patch permits to give the override directly
linebyline.

It also introduce a boolean to restart the unit after setting the
override.