mod:environment Don't modifify prompt for interactive shell
The mechanism to create histfile when there is no home available
was triggered for all non interactive commands (ssh <host> <command).

To avoid this situation setting the prompt is now bypassed when the
shell is not interactive (no 'i' in $-) (PS1 and PROMPT_COMMAND).

This also move the definition of a variable to define it before it
is used on an edge case error message.

Remove a trailing space