hiera: add priority in syslog server log files
By default, rsyslog server uses traditional syslog file format which
does not includes the priority (including the severity) and has a hard
to sort datetime format. This commit defines a specific format with
iso-formatted datetime and the priority and setup this format on
server-side. This notably makes log entries much easier to filter by
priority and to sort by datetime.