mod:kernel sysctl are all reloaded when one is added
The previous implementation was only loading the file currently
modified. This had the effect of having different files loaded in
a random order. The new method reload all files each time in the
same order as done during boot. This method ensures that proper
override rules are followed if the same setting is modified from
multiple places.