mod:gpfs Restart GPFS service after config change
The package for GPFS 4.1 do not start the gpfs service when
installed. So puppet had the opportunity to put the configuration
before starting the service. With GPFS 4.2 (on Debian) the service
is started during package install, so the puppet want touch to the
service if it is not explicitely told to restart the service when
config change.

This was the behavior before GPFS module refactoring.

Also some lint fixes.